Begin adding INDI support via LBT library.  Started with liblilxml, committing before any changes made

+++ libMagAOX/app/stateCodes.hpp
@@ -25,7 +25,7 @@ struct stateCodes
    /// The type of the state code.
    /**
      */
-   typedef int stateCodeT;
+   typedef int16_t stateCodeT;
 
 
    /// The numeric codes descrbing an application's state
